+++ include/powerloader/utils.hpp
