+++ src/context.cpp
@@ -1,4 +1,4 @@
-#include "context.hpp"
+#include <powerloader/context.hpp>
 
 #include <exception>
