+++ src/downloader.cpp
@@ -15,14 +15,14 @@
 
 namespace fs = std::filesystem;
 
-#include "context.hpp"
-#include "curl.hpp"
-#include "download_target.hpp"
-#include "downloader.hpp"
-#include "enums.hpp"
-#include "mirror.hpp"
-#include "target.hpp"
-#include "utils.hpp"
+#include <powerloader/context.hpp>
+#include <powerloader/curl.hpp>
+#include <powerloader/download_target.hpp>
+#include <powerloader/downloader.hpp>
+#include <powerloader/enums.hpp>
+#include <powerloader/mirror.hpp>
+#include <powerloader/target.hpp>
+#include <powerloader/utils.hpp>
 #ifdef WITH_ZCHUNK
 #include "zck.hpp"
 #endif
