+++ src/mirror.cpp
@@ -1,8 +1,8 @@
 #include <spdlog/fmt/fmt.h>
 
-#include "mirror.hpp"
-#include "mirrors/oci.hpp"
-#include "target.hpp"
+#include <powerloader/mirror.hpp>
+#include <powerloader/mirrors/oci.hpp>
+#include <powerloader/target.hpp>
 
 namespace powerloader
 {
