+++ src/python/main.cpp
@@ -2,7 +2,8 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <pybind11/functional.h>
-#include "downloader.hpp"
+
+#include <powerloader/downloader.hpp>
 
 namespace py = pybind11;
