+++ src/utils.cpp
@@ -1,4 +1,4 @@
-#include "utils.hpp"
+#include <powerloader/utils.hpp>
 #include <openssl/evp.h>
 
 namespace powerloader
