+++ test/test_fileio.cpp
@@ -1,7 +1,6 @@
 #include <gtest/gtest.h>
 #include <fstream>
-#include "fileio.hpp"
-#include <fstream>
+#include <powerloader/fileio.hpp>
 
 using namespace powerloader;
